+++ scm_services/__init__.py
@@ -1,7 +1,7 @@
 from scm_services.scm import SCMService, BasicSCMService
 from scm_services.adoe import ADOEService
 from scm_services.bbdc import BBDCService
-from scm_services.gh import GHServiceBasic, GHServiceChecks, GHServiceCommitStatus
 from scm_services.gl import GLService
-
-
+from scm_services.gh.checks import GHServiceChecks
+from scm_services.gh.basic import GHServiceBasic
+from scm_services.gh.commit_status import GHServiceCommitStatus
